moving verification dismissing to the robot
